examples: stream-sum refactor & E2E test cases
stream-sum is largely factored into `stream-sum` package, leaving
`chunker` and `summer` packages as minimal main()s.

A basic E2E test is added, leveraging `consumertest`.

Issue #121
